refactor: unified table in updateHistory
