+++ frontend/src/pages/main/services/ServiceTable.tsx
@@ -1,23 +1,13 @@
-import ExpandMoreIcon from "@mui/icons-material/ExpandMore";
 import PlayArrowIcon from "@mui/icons-material/PlayArrow";
 import RestartAltIcon from "@mui/icons-material/RestartAlt";
 import StopCircleIcon from "@mui/icons-material/StopCircle";
 import TerminalIcon from "@mui/icons-material/Terminal";
-import {
-  Box,
-  Table,
-  TableBody,
-  TableCell,
-  TableContainer,
-  TableHead,
-  TableRow,
-  IconButton,
-  TextField,
-  Tooltip,
-  Collapse,
-} from "@mui/material";
-import { motion } from "framer-motion";
+import { Box, TableCell, IconButton, TextField, Tooltip } from "@mui/material";
 import React, { useState } from "react";
+
+import UnifiedCollapsibleTable, {
+  UnifiedTableColumn,
+} from "@/components/tables/UnifiedCollapsibleTable";
 
 export interface Service {
   name: string;
@@ -46,13 +36,37 @@ const ServiceTable: React.FC<ServiceTableProps> = ({
   isLoading = false,
 }) => {
   const [search, setSearch] = useState("");
-  const [expanded, setExpanded] = useState<string | null>(null);
 
   const filtered = serviceList.filter(
     (s) =>
       s.name.toLowerCase().includes(search.toLowerCase()) ||
       (s.description?.toLowerCase().includes(search.toLowerCase()) ?? false),
   );
+
+  const columns: UnifiedTableColumn[] = [
+    {
+      field: "status",
+      headerName: "Status",
+      align: "left",
+      width: "120px",
+      sx: { paddingLeft: "8px" },
+    },
+    { field: "name", headerName: "Name", align: "left", width: "200px" },
+    {
+      field: "load_state",
+      headerName: "Load State",
+      align: "left",
+      width: "120px",
+    },
+    {
+      field: "sub_state",
+      headerName: "Sub State",
+      align: "left",
+      width: "120px",
+    },
+    { field: "description", headerName: "Description", align: "left" },
+    { field: "actions", headerName: "Actions", align: "right", width: "180px" },
+  ];
 
   return (
     <Box>
@@ -67,180 +81,95 @@ const ServiceTable: React.FC<ServiceTableProps> = ({
         />
         <Box fontWeight="bold">{filtered.length} shown</Box>
       </Box>
-      <TableContainer>
-        <Table size="small" sx={{ borderRadius: 3, boxShadow: 2 }}>
-          <TableHead>
-            <TableRow
-              sx={(theme) => ({
-                "& .MuiTableCell-root": { borderBottom: "none" },
-                backgroundColor:
-                  theme.palette.mode === "dark"
-                    ? "rgba(255,255,255,0.08)"
-                    : "rgba(0,0,0,0.08)",
-                borderRadius: "6px",
-                boxShadow: "none",
-              })}
-            >
-              <TableCell>Status</TableCell>
-              <TableCell>Name</TableCell>
-              <TableCell>Load State</TableCell>
-              <TableCell>Sub State</TableCell>
-              <TableCell>Description</TableCell>
-              <TableCell align="right">Actions</TableCell>
-              <TableCell />
-            </TableRow>
-          </TableHead>
-          <TableBody>
-            {filtered.map((service, index) => (
-              <React.Fragment key={service.name}>
-                <TableRow
-                  sx={(theme) => ({
-                    "& .MuiTableCell-root": { borderBottom: "none" },
-                    backgroundColor:
-                      index % 2 === 0
-                        ? "transparent"
-                        : theme.palette.mode === "dark"
-                          ? "rgba(255,255,255,0.04)"
-                          : "rgba(0,0,0,0.05)",
-                  })}
+      <UnifiedCollapsibleTable
+        data={filtered}
+        columns={columns}
+        getRowKey={(service) => service.name}
+        renderMainRow={(service) => (
+          <>
+            <TableCell sx={{ paddingLeft: "8px" }}>
+              <Box
+                component="span"
+                sx={{
+                  display: "inline-block",
+                  width: 12,
+                  height: 12,
+                  borderRadius: "50%",
+                  bgcolor:
+                    service.active_state === "active"
+                      ? "#00e676"
+                      : service.active_state === "failed"
+                        ? "#ff5252"
+                        : "#bdbdbd",
+                  mr: 1,
+                }}
+              />
+              {service.active_state}
+            </TableCell>
+            <TableCell>{service.name}</TableCell>
+            <TableCell>{service.load_state}</TableCell>
+            <TableCell>{service.sub_state}</TableCell>
+            <TableCell>{service.description || "-"}</TableCell>
+            <TableCell align="right">
+              <Tooltip title="View logs">
+                <IconButton
+                  size="small"
+                  onClick={() => onViewLogs(service)}
+                  disabled={isLoading}
                 >
-                  <TableCell>
-                    <Box
-                      component="span"
-                      sx={{
-                        display: "inline-block",
-                        width: 12,
-                        height: 12,
-                        borderRadius: "50%",
-                        bgcolor:
-                          service.active_state === "active"
-                            ? "#00e676"
-                            : service.active_state === "failed"
-                              ? "#ff5252"
-                              : "#bdbdbd",
-                        mr: 1,
-                      }}
-                    />
-                    {service.active_state}
-                  </TableCell>
-                  <TableCell>{service.name}</TableCell>
-                  <TableCell>{service.load_state}</TableCell>
-                  <TableCell>{service.sub_state}</TableCell>
-                  <TableCell>{service.description || "-"}</TableCell>
-                  <TableCell align="right">
-                    <Tooltip title="View logs">
-                      <IconButton
-                        size="small"
-                        onClick={() => onViewLogs(service)}
-                        disabled={isLoading}
-                      >
-                        <TerminalIcon fontSize="small" />
-                      </IconButton>
-                    </Tooltip>
-                    {service.active_state === "active" ? (
-                      <>
-                        <Tooltip title="Restart">
-                          <IconButton
-                            size="small"
-                            onClick={() => onRestart(service)}
-                            disabled={isLoading}
-                          >
-                            <RestartAltIcon fontSize="small" />
-                          </IconButton>
-                        </Tooltip>
-                        <Tooltip title="Stop">
-                          <IconButton
-                            size="small"
-                            onClick={() => onStop(service)}
-                            disabled={isLoading}
-                          >
-                            <StopCircleIcon fontSize="small" />
-                          </IconButton>
-                        </Tooltip>
-                      </>
-                    ) : (
-                      <Tooltip title="Start">
-                        <IconButton
-                          size="small"
-                          onClick={() => onStart(service)}
-                          disabled={isLoading}
-                        >
-                          <PlayArrowIcon fontSize="small" />
-                        </IconButton>
-                      </Tooltip>
-                    )}
-                  </TableCell>
-                  <TableCell>
+                  <TerminalIcon fontSize="small" />
+                </IconButton>
+              </Tooltip>
+              {service.active_state === "active" ? (
+                <>
+                  <Tooltip title="Restart">
                     <IconButton
                       size="small"
-                      onClick={() =>
-                        setExpanded(
-                          expanded === service.name ? null : service.name,
-                        )
-                      }
+                      onClick={() => onRestart(service)}
+                      disabled={isLoading}
                     >
-                      <ExpandMoreIcon
-                        style={{
-                          transform:
-                            expanded === service.name
-                              ? "rotate(180deg)"
-                              : "rotate(0deg)",
-                          transition: "0.2s",
-                        }}
-                      />
+                      <RestartAltIcon fontSize="small" />
                     </IconButton>
-                  </TableCell>
-                </TableRow>
-                <TableRow
-                  sx={(theme) => ({
-                    "& .MuiTableCell-root": { borderBottom: "none" },
-                    backgroundColor:
-                      index % 2 === 0
-                        ? "transparent"
-                        : theme.palette.mode === "dark"
-                          ? "rgba(255,255,255,0.08)"
-                          : "rgba(0,0,0,0.05)",
-                  })}
-                >
-                  <TableCell
-                    style={{ paddingBottom: 0, paddingTop: 0 }}
-                    colSpan={7}
-                  >
-                    <Collapse
-                      in={expanded === service.name}
-                      timeout="auto"
-                      unmountOnExit
+                  </Tooltip>
+                  <Tooltip title="Stop">
+                    <IconButton
+                      size="small"
+                      onClick={() => onStop(service)}
+                      disabled={isLoading}
                     >
-                      <Box
-                        component={motion.div}
-                        initial={{ opacity: 0, y: -10 }}
-                        animate={{ opacity: 1, y: 0 }}
-                        sx={{
-                          margin: 2,
-                          borderRadius: 2,
-                          p: 2,
-                        }}
-                      >
-                        <b>Name:</b> {service.name}
-                        <br />
-                        <b>Description:</b> {service.description || "-"}
-                        <br />
-                        <b>Load State:</b> {service.load_state}
-                        <br />
-                        <b>Active State:</b> {service.active_state}
-                        <br />
-                        <b>Sub State:</b> {service.sub_state}
-                        <br />
-                        {/* Add whatever extra info you want */}
-                      </Box>
-                    </Collapse>
-                  </TableCell>
-                </TableRow>
-              </React.Fragment>
-            ))}
-          </TableBody>
-        </Table>
-      </TableContainer>
+                      <StopCircleIcon fontSize="small" />
+                    </IconButton>
+                  </Tooltip>
+                </>
+              ) : (
+                <Tooltip title="Start">
+                  <IconButton
+                    size="small"
+                    onClick={() => onStart(service)}
+                    disabled={isLoading}
+                  >
+                    <PlayArrowIcon fontSize="small" />
+                  </IconButton>
+                </Tooltip>
+              )}
+            </TableCell>
+          </>
+        )}
+        renderExpandedContent={(service) => (
+          <>
+            <b>Name:</b> {service.name}
+            <br />
+            <b>Description:</b> {service.description || "-"}
+            <br />
+            <b>Load State:</b> {service.load_state}
+            <br />
+            <b>Active State:</b> {service.active_state}
+            <br />
+            <b>Sub State:</b> {service.sub_state}
+          </>
+        )}
+        emptyMessage="No services found."
+      />
     </Box>
   );
 };
